+++ ingest/tests.py
@@ -908,9 +908,11 @@ class IngestViewTestCase(TransactionTestCase):
 
         self.assertEqual(1, Issue.objects.count())
         self.assertEqual(1, Issue.objects.get().stored_event_count)
+        self.assertEqual(1, Project.objects.get(id=self.quiet_project.id).issue_count)
         self.assertEqual(1, Project.objects.get(id=self.quiet_project.id).stored_event_count)
 
         BaseIngestAPIView().digest_event(**_digest_params(create_event_data(), self.quiet_project, request))
+        self.assertEqual(1, Project.objects.get(id=self.quiet_project.id).issue_count)
         self.assertEqual(2, Issue.objects.get().stored_event_count)
         self.assertEqual(2, Project.objects.get(id=self.quiet_project.id).stored_event_count)